+++ tests/functional/Extension/Footnote/FootnoteExtensionGFMTest.php
@@ -1,39 +1,43 @@
 <?php
 
-declare(strict_types=1);
-
 /*
  * This file is part of the league/commonmark package.
  *
  * (c) Colin O'Dell <colinodell@gmail.com>
+ * (c) Rezo Zero / Ambroise Maupate
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace League\CommonMark\Tests\Functional\Extension\TableOfContents;
+declare(strict_types=1);
+
+namespace League\CommonMark\Tests\Functional\Extension\Footnote;
 
 use League\CommonMark\ConverterInterface;
 use League\CommonMark\Environment\Environment;
 use League\CommonMark\Extension\CommonMark\CommonMarkCoreExtension;
-use League\CommonMark\Extension\HeadingPermalink\HeadingPermalinkExtension;
-use League\CommonMark\Extension\TableOfContents\TableOfContentsExtension;
+use League\CommonMark\Extension\Footnote\FootnoteExtension;
+use League\CommonMark\Extension\GithubFlavoredMarkdownExtension;
 use League\CommonMark\MarkdownConverter;
 use League\CommonMark\Tests\Functional\AbstractLocalDataTest;
 
-final class TableOfContentsExtensionTest extends AbstractLocalDataTest
+/**
+ * Test with other extensions
+ */
+final class FootnoteExtensionGFMTest extends AbstractLocalDataTest
 {
     /**
      * @param array<string, mixed> $config
      */
     protected function createConverter(array $config = []): ConverterInterface
     {
-        $environment = new Environment($config);
-        $environment->addExtension(new CommonMarkCoreExtension());
-        $environment->addExtension(new HeadingPermalinkExtension());
-        $environment->addExtension(new TableOfContentsExtension());
+        $gfmEnvironment = new Environment($config);
+        $gfmEnvironment->addExtension(new CommonMarkCoreExtension());
+        $gfmEnvironment->addExtension(new GithubFlavoredMarkdownExtension());
+        $gfmEnvironment->addExtension(new FootnoteExtension());
 
-        return new MarkdownConverter($environment);
+        return new MarkdownConverter($gfmEnvironment);
     }
 
     /**